refactor(MemberReservationRequest): 금액 클래스 변경
- 최대 결제 금액 10억으로 설정

+++ src/main/java/roomescape/domain/PaymentInfo.java
@@ -1,18 +1,20 @@
 package roomescape.domain;
 
+import java.math.BigDecimal;
+
 public class PaymentInfo {
 
-    private final int amount;
+    private final BigDecimal amount;
     private final String orderId;
     private final String paymentKey;
 
-    public PaymentInfo(int amount, String orderId, String paymentKey) {
+    public PaymentInfo(BigDecimal amount, String orderId, String paymentKey) {
         this.amount = amount;
         this.orderId = orderId;
         this.paymentKey = paymentKey;
     }
 
-    public int getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
